Camera - Takes and displays thumbnail
If we select the 'add an image' region, the camera will now open. Once the user has taken a photo, it will save a THUMBNAIL and display it in an ImageView.
This is okay for now, but we want to save a full image and use that, as the current image is extremely blurry.