Bug fix: Fix issue where the next water string wasn't updating each day
There was a bug where the next water string & period wasn't being updated each day, it would never update at all.

Currently fixed quickly by updating both on the RecyclerView onBind, which isn't ideal as we only need to do this once each time the app opens, and this way means that each time we move back to the RecyclerView it will do the calculation again, which is redundant.

Need to refactor this but it's okay for now.